fix: e2e ldap, oauth and email specs
